+++ packages/jest-config-factory/src/next-jest.js
@@ -1,8 +1,22 @@
+import fs from "node:fs";
 import { spawn } from "node:child_process";
+import { createRequire } from "node:module";
+import yargs from "yargs";
 import findFileUpward from "./findFileUpward.js";
 
+const require = createRequire(import.meta.url);
+
+const jestCliBasePath = require
+  .resolve("jest-cli/package.json", { paths: [process.cwd()] })
+  .replace("/package.json", "");
+const jestCliArgsPath = `${jestCliBasePath}/build/args.js`;
+const { options } = require(fs.existsSync(jestCliArgsPath)
+  ? jestCliArgsPath
+  : `${jestCliBasePath}/build/cli/args.js`); // Todo: Change to import from package, when https://github.com/facebook/jest/pull/13970 is released.
+
 const args = process.argv.slice(2);
-const target = args.find((a) => !a.startsWith("-"));
+const argv = yargs(args).options(options).argv;
+const target = argv._[0];
 
 function runByLerna() {
   spawn("npx", ["lerna", "run", "test", "--", ...args], {