+++ v3/brick-kit-v3/src/index.ts
@@ -1,0 +1,9 @@
+import antdLocaleZhCN from "antd/es/locale/zh_CN";
+
+export * from "../../../packages/brick-kit/src/decorators";
+export * from "../../../packages/brick-kit/src/UpdatingElement";
+
+export {
+  // This is not in dll manifest.
+  antdLocaleZhCN,
+};