+++ packages/build-config-factory/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.25.2](https://github.com/easyops-cn/next-core/compare/@next-core/build-config-factory@2.25.1...@next-core/build-config-factory@2.25.2) (2023-03-02)
+
+**Note:** Version bump only for package @next-core/build-config-factory
+
+## [2.25.1](https://github.com/easyops-cn/next-core/compare/@next-core/build-config-factory@2.25.0...@next-core/build-config-factory@2.25.1) (2023-03-02)
+
+**Note:** Version bump only for package @next-core/build-config-factory
+
+# [2.25.0](https://github.com/easyops-cn/next-core/compare/@next-core/build-config-factory@2.24.7...@next-core/build-config-factory@2.25.0) (2023-02-28)
+
+### Features
+
+- properties doc support editor ([ec90293](https://github.com/easyops-cn/next-core/commit/ec90293be6cd7d0cf0efd934edfbbeaa0b085dbf))
+
+## [2.24.7](https://github.com/easyops-cn/next-core/compare/@next-core/build-config-factory@2.24.6...@next-core/build-config-factory@2.24.7) (2023-02-20)
+
+**Note:** Version bump only for package @next-core/build-config-factory
+
 ## [2.24.6](https://github.com/easyops-cn/next-core/compare/@next-core/build-config-factory@2.24.5...@next-core/build-config-factory@2.24.6) (2023-02-16)
 
 **Note:** Version bump only for package @next-core/build-config-factory