+++ packages/next-docs/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.431](https://github.com/easyops-cn/next-core/compare/@next-core/next-docs@1.2.430...@next-core/next-docs@1.2.431) (2023-03-09)
+
+**Note:** Version bump only for package @next-core/next-docs
+
+## [1.2.430](https://github.com/easyops-cn/next-core/compare/@next-core/next-docs@1.2.429...@next-core/next-docs@1.2.430) (2023-03-09)
+
+**Note:** Version bump only for package @next-core/next-docs
+
+## [1.2.429](https://github.com/easyops-cn/next-core/compare/@next-core/next-docs@1.2.428...@next-core/next-docs@1.2.429) (2023-03-03)
+
+**Note:** Version bump only for package @next-core/next-docs
+
+## [1.2.428](https://github.com/easyops-cn/next-core/compare/@next-core/next-docs@1.2.427...@next-core/next-docs@1.2.428) (2023-03-02)
+
+**Note:** Version bump only for package @next-core/next-docs
+
+## [1.2.427](https://github.com/easyops-cn/next-core/compare/@next-core/next-docs@1.2.426...@next-core/next-docs@1.2.427) (2023-02-20)
+
+**Note:** Version bump only for package @next-core/next-docs
+
 ## [1.2.426](https://github.com/easyops-cn/next-core/compare/@next-core/next-docs@1.2.425...@next-core/next-docs@1.2.426) (2023-02-16)
 
 **Note:** Version bump only for package @next-core/next-docs